requirements.txt
boto3
pillow
keras
h5py
pandas
tensorflow
scikit-image
